small renaming, semantic task in beginning
